chore(avm): rename avm_inputs to avm_io
