+++ barretenberg/cpp/src/barretenberg/vm2/common/avm_io.hpp
