+++ barretenberg/cpp/src/barretenberg/vm2/constraining/relations/tx.test.cpp
@@ -4,7 +4,7 @@
 #include <cstdint>
 #include <vector>
 
-#include "barretenberg/vm2/common/avm_inputs.hpp"
+#include "barretenberg/vm2/common/avm_io.hpp"
 #include "barretenberg/vm2/common/aztec_constants.hpp"
 #include "barretenberg/vm2/common/aztec_types.hpp"
 #include "barretenberg/vm2/constraining/flavor_settings.hpp"
