+++ barretenberg/cpp/src/barretenberg/vm2/proving_helper.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "barretenberg/honk/proof_system/types/proof.hpp"
-#include "barretenberg/vm2/common/avm_inputs.hpp"
+#include "barretenberg/vm2/common/avm_io.hpp"
 #include "barretenberg/vm2/constraining/prover.hpp"
 #include "barretenberg/vm2/constraining/verifier.hpp"
 #include "barretenberg/vm2/tracegen/trace_container.hpp"
