+++ barretenberg/cpp/src/barretenberg/vm2/simulation/events/tx_events.hpp
@@ -3,7 +3,7 @@
 #include <cstdint>
 #include <variant>
 
-#include "barretenberg/vm2/common/avm_inputs.hpp"
+#include "barretenberg/vm2/common/avm_io.hpp"
 #include "barretenberg/vm2/common/aztec_types.hpp"
 #include "barretenberg/vm2/common/field.hpp"
 #include "barretenberg/vm2/simulation/events/tx_context_event.hpp"
