+++ barretenberg/cpp/src/barretenberg/vm2/simulation/gadgets/nullifier_tree_check.cpp
@@ -1,6 +1,6 @@
 #include "barretenberg/vm2/simulation/gadgets/nullifier_tree_check.hpp"
 
-#include "barretenberg/vm2/common/avm_inputs.hpp"
+#include "barretenberg/vm2/common/avm_io.hpp"
 #include "barretenberg/vm2/common/constants.hpp"
 
 namespace bb::avm2::simulation {
