+++ barretenberg/cpp/src/barretenberg/vm2/simulation/gadgets/retrieved_bytecodes_tree_check.test.cpp
@@ -4,7 +4,7 @@
 #include <gtest/gtest.h>
 
 #include "barretenberg/crypto/merkle_tree/memory_tree.hpp"
-#include "barretenberg/vm2/common/avm_inputs.hpp"
+#include "barretenberg/vm2/common/avm_io.hpp"
 #include "barretenberg/vm2/common/aztec_types.hpp"
 #include "barretenberg/vm2/simulation/events/retrieved_bytecodes_tree_check_event.hpp"
 #include "barretenberg/vm2/simulation/lib/merkle.hpp"
