+++ barretenberg/cpp/src/barretenberg/vm2/testing/fixtures.hpp
@@ -3,7 +3,7 @@
 #include <vector>
 
 #include "barretenberg/crypto/merkle_tree/memory_tree.hpp"
-#include "barretenberg/vm2/common/avm_inputs.hpp"
+#include "barretenberg/vm2/common/avm_io.hpp"
 #include "barretenberg/vm2/common/aztec_types.hpp"
 #include "barretenberg/vm2/common/field.hpp"
 #include "barretenberg/vm2/simulation/lib/serialization.hpp"
