+++ barretenberg/cpp/src/barretenberg/vm2/testing/public_inputs_builder.hpp
@@ -4,7 +4,7 @@
 #include <vector>
 
 #include "barretenberg/numeric/random/engine.hpp"
-#include "barretenberg/vm2/common/avm_inputs.hpp"
+#include "barretenberg/vm2/common/avm_io.hpp"
 #include "barretenberg/vm2/common/aztec_types.hpp"
 #include "barretenberg/vm2/common/field.hpp"
 #include "barretenberg/vm2/simulation/lib/serialization.hpp"
