+++ barretenberg/cpp/src/barretenberg/vm2/tracegen/nullifier_tree_check_trace.test.cpp
@@ -6,7 +6,7 @@
 #include <gtest/gtest.h>
 
 #include "barretenberg/crypto/poseidon2/poseidon2.hpp"
-#include "barretenberg/vm2/common/avm_inputs.hpp"
+#include "barretenberg/vm2/common/avm_io.hpp"
 #include "barretenberg/vm2/constraining/flavor_settings.hpp"
 #include "barretenberg/vm2/constraining/testing/check_relation.hpp"
 #include "barretenberg/vm2/generated/relations/lookups_nullifier_check.hpp"
